% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{periodicTable}
\alias{periodicTable}
\title{Periodic table of elements}
\format{
A data frame with 118 rows and 28 columns.
See documentation in https://gist.github.com/GoodmanSciences/c2dd862cd38f21b0ad36b8f96b4bf1ee
}
\source{
\url{https://gist.github.com/GoodmanSciences/c2dd862cd38f21b0ad36b8f96b4bf1ee}
}
\usage{
periodicTable
}
\description{
Properties of chemical elements
}
\keyword{datasets}
