% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{petroOxides}
\alias{petroOxides}
\title{Oxides of petrological interest}
\format{
A data frame with 22 rows and 6 columns.
\describe{
\item{OxideName}{Full formula of the oxide}
\item{CationName}{}
\item{NbCations, NbO}{Number of cations and oxygen atoms in the oxide}
\item{CationMass}{Mass of the cation, taken from GCDmodel::periodicTable}
\item{OxideMass}{Mass of the oxide, calculated from the above data}
}
}
\source{
J.F. Moyen
}
\usage{
petroOxides
}
\description{
Properties of chemical elements
}
\keyword{datasets}
